#177 delete node comments fix up
